+++ packages/datum/test/core/datum_metrics_test.dart
@@ -129,7 +129,7 @@ void main() {
         () => localAdapter.saveLastSyncResult(any(), any()),
       ).thenAnswer((_) async {});
 
-      datum = (await Datum.initialize(
+      (await Datum.initialize(
         config: const DatumConfig(schemaVersion: 0),
         connectivityChecker: connectivityChecker,
         registrations: [
@@ -139,7 +139,10 @@ void main() {
           ),
         ],
       ))
-          .getSuccess();
+          .fold(
+        (l, s) => throw l,
+        (r) => datum = r,
+      );
     });
 
     tearDown(() async {